updated scripts to be able to wait to run

+++ test-scripts/initial.sh
@@ -1,23 +1,33 @@
-#!/usr/bin/sh
+#!/usr/bin/bash
 
-echo "ping 8.8.8.8 -c 10 ----------------------------------------------"
-ping 8.8.8.8 -c 10
-sleep 3
-echo "wget google.com -------------------------------------------------"
-wget google.com
-sleep 15
-echo "wget www.flyn.org -----------------------------------------------"
-wget www.flyn.org
-sleep 15
-echo "wget www.flyn.org/courses/ --------------------------------------"
-wget www.flyn.org/courses/
-sleep 15
-echo "wget www.aquinas.dev --------------------------------------------"
-wget www.aquinas.dev
-sleep 15
-echo "ping google.com -c 10 -------------------------------------------"
-ping google.com -c 10
-sleep 5
-echo "cleaning up -----------------------------------------------------"
-rm -f index.html*
-echo "FINISHED SCRIPT 1"
+echo "+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
+echo "Initial Script - Diverse checks for simple graph check"
+echo "begin script? (y/n):"
+read input
+if [ $input != 'y' ]; then
+    echo "script skipped"
+    exit 0
+else
+    echo "+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
+    ping 8.8.8.8 -c 10
+    sleep 3
+    echo "wget google.com -------------------------------------------------"
+    wget google.com
+    sleep 15
+    echo "wget www.flyn.org -----------------------------------------------"
+    wget www.flyn.org
+    sleep 15
+    echo "wget www.flyn.org/courses/ --------------------------------------"
+    wget www.flyn.org/courses/
+    sleep 15
+    echo "wget www.aquinas.dev --------------------------------------------"
+    wget www.aquinas.dev
+    sleep 15
+    echo "ping google.com -c 10 -------------------------------------------"
+    ping google.com -c 10
+    sleep 5
+    echo "cleaning up -----------------------------------------------------"
+    echo "rm -f index.html*"
+    rm -f index.html*
+    echo "FINISHED INITIAL SCRIPT -----------------------------------------"
+fi
